+++ HISTORY.rst
@@ -24,6 +24,7 @@ latest (unreleased)
 * Add 'client_order_ref' in SO Tree view
 * Install account_invoice_rounding & account_bank_statement_import_camt
 * Activate multi-company features
+* Load custom chart of accounts
 
 **Bugfixes**